Simplify SensorWorker by keeping only one implementation
