+++ wear/src/main/kotlin/io/homeassistant/companion/android/home/HomeActivity.kt
@@ -16,6 +16,7 @@ import androidx.lifecycle.lifecycleScope
 import androidx.lifecycle.repeatOnLifecycle
 import androidx.wear.protolayout.ActionBuilders
 import dagger.hilt.android.AndroidEntryPoint
+import io.homeassistant.companion.android.common.sensors.SensorWorker
 import io.homeassistant.companion.android.common.util.SdkVersion
 import io.homeassistant.companion.android.home.views.DEEPLINK_PREFIX_SET_CAMERA_TILE
 import io.homeassistant.companion.android.home.views.DEEPLINK_PREFIX_SET_SHORTCUT_TILE
@@ -24,7 +25,6 @@ import io.homeassistant.companion.android.home.views.DEEPLINK_PREFIX_SET_THERMOS
 import io.homeassistant.companion.android.home.views.LoadHomePage
 import io.homeassistant.companion.android.onboarding.OnboardingActivity
 import io.homeassistant.companion.android.sensors.SensorReceiver
-import io.homeassistant.companion.android.sensors.SensorWorker
 import io.homeassistant.companion.android.tiles.OpenTileSettingsActivity
 import javax.inject.Inject
 import kotlinx.coroutines.CancellationException
